Add indexing and simple searching possibility
The Searchable module is added which is included in both User and
Product models. It integrates Elasticsearch with models by including
the Elasticsearch::Model main module. It also adds the after_commit
callback, which indexes data after every operation.

To speed importing data for the first time, the
ElasticsearchDataImporter is used. It firstly creates new index for
given model, prepares data and then uses bulk index API to import
data in batches. It can be really helpful when adding Elasticsearch
integration to existing project with large dataset.