Add multi-model search functionality
We now can search both User and House models by exact match of any
field. It uses the Autocompleter service object which simply maps
returned list of records to specific format. We need a record type,
record id and some record hint to know which record we want to
select. Multi-model search is possible with the search method
called on Elasticsearch::Model.

The search in run after 0.5 second with use of the "debounce"
function. It simply waits given time before executing function.

Also look at this ugly method of building hints in the Autocompleter.
As a bonus we will refactor it in the future to use polymorphism.