Add custom search query
When searching the query is now treated as a one big query with the
"operator: and" option. The default is "or" so every word could return
its own results. It can be misleading and confusing.

Also if both User and House match given query the House is always
returned first, it is accomplished by adding a big weight for house
type.

Searching should return 50 results.